+++ src/pine/parser.clj
@@ -68,17 +68,31 @@
 
 (defn- -normalize-column [column]
   (match column
-    [:column [:symbol c]]                                   {:column c}
-    [:column [:symbol a] [:symbol c]]                       {:alias a :column c}
-    [:column [:symbol c]]                                   {:column c}
-    [:column [:symbol c] [:alias [:symbol ca]]]             {:column c :column-alias ca}
-    [:column [:symbol a] [:symbol c] [:alias [:symbol ca]]] {:alias a :column c :column-alias ca}
+    [:aliased-column [:column [:symbol c]]]                                     {:column c}
+    [:aliased-column [:column [:alias [:symbol a]] [:symbol c]]]                {:alias a :column c}
+    [:aliased-column [:column [:symbol c]] [:alias [:symbol ca]]]             {:column c :column-alias ca}
+    [:aliased-column [:column [:alias [:symbol a]] [:symbol c]] [:alias [:symbol ca]]] {:alias a :column c :column-alias ca}
 
     :else                 (throw (ex-info "Unknown COLUMN operation" {:_ column}))))
 
 (defmethod -normalize-op :SELECT [[_ payload]]
   (match payload
-    [:columns & columns] {:type :select :value (mapv -normalize-column columns)}
+    [:aliased-columns & columns] {:type :select :value (mapv -normalize-column columns)}
+    :else                (throw (ex-info "Unknown SELECT operation" {:_ payload}))))
+
+;; ------
+;; ORDER
+;; ------
+
+(defn- -normalize-order [column]
+  (match column
+    [:order-column [:column [:symbol c]]]   {:column c :direction "DESC"}
+    [:order-column [:column [:symbol c]] d] {:column c :direction (clojure.string/upper-case d)}
+    :else                                   (throw (ex-info "Unknown ORDER operation" {:_ column}))))
+
+(defmethod -normalize-op :ORDER [[_ payload]]
+  (match payload
+    [:order-columns & columns] {:type :order :value (mapv -normalize-order columns)}
     :else                (throw (ex-info "Unknown SELECT operation" {:_ payload}))))
 
 ;; -----