+++ test/pine/ast_test.clj
@@ -52,6 +52,13 @@
     (is (= []
            (generate :columns "user"))))
 
+  (testing "Generate ast for `order`"
+    (is (= [{:alias "c_0" :column "country" :direction "DESC"}]
+           (generate :order "company | o: country")))
+    (is (= [{:alias "c_0" :column "country" :direction "DESC"}
+            {:alias "c_0" :column "created_at" :direction "DESC"}]
+           (generate :order "company | o: country, created_at"))))
+
   (testing "Generate ast for `limit`"
     (is (= 10
            (generate :limit "limit: 10")))