+++ test/pine/parser_test.clj
@@ -84,5 +84,12 @@
     (is (= [{:type :where, :value ["age" "IN" [(dt/string "24") (dt/string "36")]]}] (p "age in (  '24' ,'36' )")))
     (is (= [{:type :where, :value ["age" "IN" [(dt/string "24") (dt/string "36")]]}] (p "age in ('24' '36')"))))
 
+  (testing "Parse `order` expressions"
+    (is (= [{:type :order, :value [{:column  "name" :direction "DESC"}]}]            (p "order: name")))
+    (is (= [{:type :order, :value [{:column  "name" :direction "DESC"}]}]            (p "order: name desc")))
+    (is (= [{:type :order, :value [{:column  "name" :direction "ASC"}]}]            (p "order: name asc")))
+    (is (= [{:type :order, :value [{:column  "name" :direction "DESC"}
+                                   {:column "created_at" :direction "ASC"}]}]       (p "order: name, created_at asc"))))
+
   (testing "Parse `delete` expressions"
     (is (= [{:type :delete, :value {:column "id"}}] (p "delete! .id")))))